+++ common/src/main/java/dev/terminalmc/chatnotify/command/Commands.java
@@ -18,15 +18,46 @@ package dev.terminalmc.chatnotify.command;
 
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.CommandDispatcher;
+import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import dev.terminalmc.chatnotify.ChatNotify;
 import dev.terminalmc.chatnotify.gui.screen.RootScreen;
+import dev.terminalmc.chatnotify.util.Unicode;
 import net.minecraft.client.Minecraft;
 import net.minecraft.commands.CommandBuildContext;
+import net.minecraft.network.chat.Component;
 
+import java.util.List;
+
+import static net.minecraft.commands.Commands.argument;
 import static net.minecraft.commands.Commands.literal;
 
 public class Commands {
+
+    public static final List<String> FORMAT_CODES = List.of(
+            "0",
+            "1",
+            "2",
+            "3",
+            "4",
+            "5",
+            "6",
+            "7",
+            "8",
+            "9",
+            "a",
+            "b",
+            "c",
+            "d",
+            "e",
+            "f",
+            "l",
+            "m",
+            "n",
+            "o",
+            "k",
+            "r"
+    );
 
     private Commands() {
         throw new UnsupportedOperationException("This class cannot be instantiated.");
@@ -40,6 +71,27 @@ public class Commands {
                     mc.tell(() -> mc.setScreen(new RootScreen(mc.screen)));
                     return Command.SINGLE_SUCCESS;
                 })
+                .then(literal("format")
+                        .then(argument("pattern", StringArgumentType.greedyString())
+                                .suggests(((ctx, builder) -> {
+                                    String[] split = ctx.getInput().split("format ");
+                                    String input = split.length < 2 ? "" : split[1];
+                                    String pre = input.endsWith("$") ? input : input + "$";
+                                    FORMAT_CODES.forEach((s) -> builder.suggest(pre + s));
+                                    return builder.buildFuture();
+                                }))
+                                .executes(ctx -> {
+                                    String pattern = StringArgumentType.getString(ctx, "pattern");
+                                    Component text = Component.literal(pattern.replaceAll(
+                                            "\\$",
+                                            Unicode.SECTION.str
+                                    ));
+
+                                    mc.gui.getChat().addMessage(text);
+                                    return Command.SINGLE_SUCCESS;
+                                })
+                        )
+                )
         );
     }
 }