+++ common/src/main/java/dev/terminalmc/chatnotify/gui/widget/list/root/ControlList.java
@@ -143,5 +143,6 @@ public class ControlList extends OptionList {
                         ));
             }
         }
+
     }
 }